% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annotated_heatmap}
\alias{annotated_heatmap}
\title{Plot an heatmap}
\usage{
annotated_heatmap(geneSel, heatmap_colors, name1, name2, metric, clust_validity,
  clust_stability, png_name)
}
\arguments{
\item{geneSel}{The selected genes}

\item{heatmap_colors}{Colors to use for the heatmap}

\item{name1}{Name of the desired state samples of the signature}

\item{name2}{Name of the control samples of the signature}

\item{metric}{Distance metric (euclidian or correlation)}

\item{clust_validity}{Cluster validity scores}

\item{f}{clust_stability cluster stability scores}

\item{png_name}{name of the png to write}
}
\description{
Plot an heatmap
}

