% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_mapping}
\alias{check_mapping}
\title{Check how many zebrafish affymatrix probes are mapped to zebrafish genes}
\usage{
check_mapping(conversion_table, outfolder_location)
}
\arguments{
\item{microarray_data}{The microarray input data}

\item{control_desired_factor}{Factor outlining which samples are control and which are desired expression state}
}
\description{
Checks the loaded conversion table for how well the conversion is done, and by which
source the conversion is done.
}

