% Generated by roxygen2 (4.0.1): do not edit by hand
\name{differential}
\alias{differential}
\title{Function to tell GOdata from which score genes are differential}
\usage{
differential(allScore)
}
\arguments{
\item{allScore}{A vector with 1's for signature genes and 0's for non-signature genes}
}
\value{
Boolean vector
}
\description{
Function to tell GOdata from which score genes are differential
}

