% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enrichment}
\alias{enrichment}
\title{Find the enriched GO terms for given genes}
\usage{
enrichment(all_genes, outfolder, signature_name, pvalue = 0.05)
}
\arguments{
\item{all_genes}{Named vector where names are genes, and value is 1 if it's part of signature, 0 if it's not}

\item{outfolder}{Outfolder to write the results to}

\item{siganture_name}{Name of the signature}

\item{pvalue}{p-value at which to consider an enriched GO term significant}
}
\description{
Takes human genes and finds the enriched GO terms for the signature. Additionally,
printes extra information about the enriched GO terms
}

