% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enrichment_overlap}
\alias{enrichment_overlap}
\title{Plot overlap in enriched GO terms between the signatures}
\usage{
enrichment_overlap(enrichment_list, ylab = "", colnames = TRUE,
  outfolder_location, GO_results_name, main_extra_info = "")
}
\arguments{
\item{enrichment_list}{A list of signatures and its GO terms}

\item{ylab}{y-axis description for the enrichment overlap plot (default: '')}

\item{colnames}{Boolean, if TRUE plot the names of the signatures on the x-axis}

\item{outfolder_location}{Location of the folder where to write the plots}

\item{GO_results_name}{Name of the subfolder to write the results in}

\item{main_extra_info}{Extra info to put in the graph title (default: '')}
}
\description{
Takes human genes and finds the enriched GO terms for the signature. Additionally,
printes extra information about the enriched GO terms
}

