% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_detailed_connectivity_results}
\alias{plot_detailed_connectivity_results}
\title{Plot results from the detailed connectivity results page}
\usage{
plot_detailed_connectivity_results(connectivity_data_frame, outfolder,
  signature_name)
}
\description{
<todo>
}

