% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_results}
\alias{plot_results}
\title{Plot information on the signatures}
\usage{
plot_results(name1, name2, geneSel, heatmap_location,
  dendogram_location_samples, dendogram_location_genes, volcano_location, f,
  fit)
}
\arguments{
\item{name1}{Name of the desired state samples of the signature}

\item{name2}{Name of the control samples of the signature}

\item{geneSel}{The selected genes}

\item{heatmap_location}{Folder location to write heatmap}

\item{dendogram_location_samples}{Folder location to write dendogram}

\item{volcano_location}{Folder location to write volcano plots}

\item{f}{Definitition of which sample is control and which is 'treatment'}

\item{fit}{The fitting object from when the differntial genes were selected}
}
\description{
Plot a lot of information on the signatures. This includes heatmaps (euclidian and correlation distance),
sample and gene dendograms and volcano plots
}

