% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validity}
\alias{validity}
\title{Calculate the validity of the clustering of samples}
\usage{
validity(geneSel, metric)
}
\arguments{
\item{geneSel}{The selected genes}

\item{metric}{The metric to use for clustering (see clValid documentation for all the options)}
}
\description{
For each signature a heatmap, among other results, is created. In the heatmap the samples are clustered
according to their expression profiles. This calculates the validity of that clustering
}

