% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{zebrafish_conversion_table}
\alias{find_drugs}
\alias{zebrafish_conversion_table}
\title{Find candidate drugs}
\usage{
load(zebrafish_conversion_table)
}
\arguments{
\item{microarray_data}{The microarray input data. A dataframe with an ID column, Gene.Symbol column, gene_assignment collumn, and N amount of columns for samples, with normalized expression values. (See microarray_data in drugFinder library)}

\item{datasets}{Datasets of}

\item{outfolder_location_root}{Path to folder where all results are written to (if it doesn't exist, will be created)}

\item{memory_limit}{Memory limit to set to R (only works in Windows)}
}
\description{
Takes in a zebrafish microarray dataset and which samples are control and desired,
and uses this to find design signatures and find top drug candidates for inducing
these signatures.
}
\keyword{datasets}

